requirements.txt
beautifulsoup4==4.11.2
python-dotenv==0.21.1
requests==2.28.1
spotipy==2.22.1
